fixed bug with dec_to_hex that caused it to cut short large numbers

+++ includes/asm.h
@@ -6,7 +6,7 @@
 /*   By: osalmine <osalmine@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/09 17:29:14 by cphillip          #+#    #+#             */
-/*   Updated: 2021/02/15 16:54:07 by osalmine         ###   ########.fr       */
+/*   Updated: 2021/02/15 22:39:15 by osalmine         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -96,7 +96,7 @@ uint8_t				*split_hex(int32_t nb, int nb_len);
 void				set_bytes(t_master *m, uint8_t *bytes, \
 						int bytes_amount, int mem_offset);
 void				set_a_byte(t_master *m, uint8_t byte, int mem_offset);
-uint16_t			dec_to_hex(int nb);
+uint32_t			dec_to_hex(int nb);
 uint8_t				*ascii_to_hex(char *str);
 void				encode_ins(t_master *m);
 void				write_to_file(t_master *m);
